Enable  show scope option in Blipp plugin
We use the [blipp](https://github.com/danielb2/blipp) to display our routes when the service starts up. As we are now using [Route scopes](https://hapi.dev/api/?v=20.0.0#-routeoptionsauthaccessscope) as part of our authorisation solution being able to see the scopes in the Blipp output will be handy.

This change enables the scope output and expands the current documentation for the plugin.