Testing if sonarcloud honours unused param convention
There is a convention in JavaScript where if a method has a particular signature but you do not use the params, you can prefix the param with an underscore to confirm you do not intend to use it.

VSCode with standardjs seems to recognise it. This test is about seeing if SonarCloud does to. Ideally if a method is expected to have a particular signature, for example, any Hapi handlers we define we would rather keep the param visible even if its not used.