Merge airbrake endpoints and add description
We have agreed a convention in the team to put any checks or tests of an environment on the `/status` path. So the first change is moving the endpoint there.

Also checking airbrake is something we should only need to do once and it would be cleaner to do in a single hit (the result is the). So this change merges the endpoints.

Finally, we add a description to the route if only so we can note that an error is expected.