Add disinfect plugin to santize inputs to the API (#4)
https://trello.com/c/cjqqnLGs

During PEN testing of the [charging-module-api](https://github.com/DEFRA/charging-mdoule-api) it was found that the project is susceptible to XSS. This was because inputs to the API were not properly being sanitized. Rather than roll our own sanitizer, we're opting to use a plugin on this project.

[Disinfect](https://github.com/genediazjr/disinfect) has been used on a number of Defra projects. It applies Google's [Caja](https://github.com/google/caja) HTML Sanitizer on route query, payload, and params.

This change adds it to this project.