Add Airbrake error logging to the project (#6)
It's a standard across Defra projects to use the relevant [Airbrake package](https://www.npmjs.com/package/@airbrake/node) to capture errors within our apps and automatically record them in an instance of [Errbit](https://github.com/errbit/errbit) we maintain. Doing so means teams can monitor and track issues in our apps without resorting to scanning logs for issues or waiting for calls from users.

This project is no exception. So this change adds support for Airbrake and includes a controller that allows us to test that airbrake is working correctly within an environment.

It also covers our first attempts at testing against an endpoint which requires authorisation. Hence the work with the authorisation test helpers and the start of a README covering guidance and notes about our tests.