Add dependabot.yml to repo
Normally for our projects, we would integrate them with <https://dependabot.com>. But we have known for a while now that [Dependabot is joining GitHub](https://dependabot.com/blog/hello-github/) and that dependency updates would one day be done natively in GitHub.

That day [seems to have come](https://github.blog/2020-06-01-keep-all-your-packages-up-to-date-with-dependabot/). Though flagged as `beta` it seems you can now keep your dependencies updated without having to add the `dependabot-preview` integration.

We already have dependabot security updates setup configured for this project. This is our first attempt to hook a project up to the native GitHub general dependency updates.