Update dependabot.yml
Having added the config needed to enable native Dependabot support we spotted that for those projects still using **Dependabot-preview** there is an option in the dashboard to switch to native. When you do that it creates a config file formatted like this

```yml
version: 2
updates:
- package-ecosystem: npm
  directory: "/"
  schedule:
    interval: daily
  open-pull-requests-limit: 10
  versioning-strategy: lockfile-only
```

As we have all our projects configured the same way in **Dependabot-preview** and that works nicely for us, this change updates what we have done here to be consistent.