Fix airbrake controller test
We recently started work on migrating [authorised systems](https://github.com/DEFRA/sroc-charging-module-api/pull/7). The old charging module determines the admin user by looking at an env var. All other users are in the DB table.

We want to be able to audit who has been making changes and link them to a user (authorised system) record in the DB. So to enable that ability in the SROC version of the API the admin user will be a record in the DB. We will just use the env var as part of seeding the db with the admin user.

What we hadn't spotted is that change has actually broken the existing airbrake controller test. It still appears to be working because it is returning a 500. But on closer inspection our code in the controller is not actually firing because the error is happening in the jwt auth strategy code.

This change updates the airbrake controller test to get it working as expected. It will also include seeding and cleaning the db as part of testing.