Fix airbrake controller test (#15)
We recently started work on migrating [authorised systems](https://github.com/DEFRA/sroc-charging-module-api/pull/7). The old charging module determines the admin user by looking at an env var. All other users are in the DB table.

We want to be able to audit who has been making changes and link them to a user (authorised system) record in the DB. So to enable that ability in the SROC version of the API the admin user will be a record in the DB. We will just use the env var as part of seeding the DB with the admin user.

What we hadn't spotted is that change has actually broken the existing airbrake controller test. It still appears to be working because it is returning a 500. But on closer inspection, our code in the controller is not actually firing because the error is happening in the JWT auth strategy code.

This change updates the airbrake controller test to get it working as expected. This mainly involved adding support for cleaning and seeding the DB as part of testing.

Things we did

- Update `db/index.js` to include Knex config
- Add database test helper with 2 functions; clean the database, and create the admin user

We anticipate a number of tests will want to call both these commands in `beforeEach` blocks. We need an admin authorised system else authentication will not work. And it's best practise to ensure the DB is empty before starting any test. You know then it contains only what is needed for your test.

On top of that, we did some housekeeping to improve requiring things. This change includes adding some `index.js` files in certain folders so that when we need them, we can `require()` the folder and use [destructuring](https://javascript.info/destructuring-assignment) to pull out the things we actually need