Add a new route helper
With a change to the authentication its about time we created some tests for it. In preparation for that this change adds a new route helper which we can use to dynamically add routes to the server during testing.

This means we don't have to rely on existing routes when testing generic functionality like authentication.