Split client creation and db clean helpers
We need to add an ability to create a non admin authorised system (client). We could have added this to the existing database helper. But we can already see a pattern of just adding anything to do with the database into this one helper.

So to keep our helpers small and focused this change splits out the existing helper. It retains the generic functionality of cleaning the database. But anything to do with 'client' systems is now in its own helper.