Update authorisation helper
We need to add an ability to create a JWT which encodes a client_id that is not the admin users. So this change adds that method.

We are also going to need to stub the JsonWebToken.verify method in scenarios where a token is expired or the signature is invalid. In those cases it throws it's own errors that hapi-now-auth is relying on to determine whether to authenicate a request.