Add authentication feature test
First, the use of the term `feature`. Authentication isn't something specific to a plugin, hook, model etc. It's a 'feature' of the service itself. As such tests for this generic features if the service need a home and 'features' seems appropriate. For example its a term used as part of BDD and tools like Cucumber.

This adds our first authentication feature test which aims to cover all the aspects of authenticating with the API including expired JWTs, incorrectly signed JWTs, and public and non-public paths.

N.B. We don't currently have anything but admin and public paths. We'll cover non-admin/non-public paths when we finally implement one!