Decide to stop fighting the name
AWS Cognito refers to them as clients hence the term client ID everywhere. But our table and end point uses the term 'authorised system'.

That isn't going to change and we are likely to add helpers for the other models based on their names. So doesn't make sense to have an exception for authorised system where we use the name 'client' instead.

So this updates the helper to make it consistent with what we expect we are going to need.