Enable JWT token expiry checking (#12)
All JWT tokens have an expiry value. [Until recently](https://aws.amazon.com/about-aws/whats-new/2020/08/amazon-cognito-user-pools-supports-customization-of-token-expiration/) those generated by AWS Cognito all had an expiry of 1 hour.

The idea is to protect against leaked tokens. Should one become compromised you want to limit how long it has access before it is required to refresh.

In the current [charging-module-api](https://github.com/defra/charging-module-api) we just decode the token. In this project at least we are verifying it. But we are still ignoring the expiry value.

This change will close that final security hole and update our auth strategy config to include checking the expiry value of tokens. To keep local development tolerable though we'll drive it through configuration. So locally we'll continue to ignore the expiry value but when deployed to our environments all tokens will be verified and checked properly.

Note

With the final step done for our authentication we also took this opportunity to add some testing for it. This change includes the additions to the test helpers to support this. It also includes the concept of 'feature tests'. Authentication isn't something specific to a plugin, hook, model etc. It's a 'feature' of the service itself. As such tests for this generic feature needed a home and 'features' seems appropriate. For example, its a term used as part of BDD and tools like Cucumber.

You'll also see us start to use JSDoc comments in our files. This is something we want to expand on as we progress work on the API.