Move payload check to correct extension point
When I added the unit tests they were failing (as expected) but not for the right reason. They kept returning a HTTP 400 Bad request. With some handy console.logging I tracked it down to our `on_request.hook.js` which was throwing the error.

It was reporting that the payload was always empty. I'm sure we had this working! 🤷 This is the first time we are making a post request in test. Plus a quick review of the Hapi lifecycle does seem to suggest that `payload` is only added to the request in the pre-auth stage https://futurestud.io/downloads/hapi/request-lifecycle

So to document and make it clear why this change has been done, this commit just deals with moving the extension point our OnRequestHook hooks on to.

We'll deal with the naming in a subsequent commit.