Add Unescape sanitized HTML characters plugin (#22)
https://trello.com/c/cjqqnLGs
https://github.com/DEFRA/sroc-charging-module-api/pull/4

We added [Disinfect](https://github.com/genediazjr/disinfect) in PR #4 to deal with the fact we were susceptible to XSS attacks.

Disinfect not only deals with malicious content, but it also escapes all content to make it safe for displaying in a browser. This means characters like `&`, `<` and `>` also get escaped. We're not worried about these in our data and feel it's for our clients to determine how to handle them in our responses.

So this change adds a new custom plugin that will unescape payloads after disinfect has escaped them.

As part of this change, we also had to move the payload check to the 'correct' extension point.

When we added the unit tests they were failing (as expected) but not for the right reason. They kept returning a `HTTP 400 Bad request`. With some handy console logging we tracked it down to our `on_request.hook.js` which was throwing the error.

It was reporting that the payload was always empty. We're sure we had this working previously! 🤷 However, this was the first time we are making a POST request in test. Plus a quick review of the Hapi lifecycle does seem to suggest that `payload` is only added to the request in the pre-auth stage https://futurestud.io/downloads/hapi/request-lifecycle

The hook It is incorrectly named after this change. But we intend to move it to a plugin in a subsequent commit anyway so have done the minimum to avoid any confusion.