Add tests for features
We have a test of the feature which enables logging in the test environment when enabled through an env var.

The issue is it will only work if

- it's the only test run
- it's the first test run

What we can't understand is how with the spy completely isolated it works in isolation but fails when run with anything else.

So we've had to makr it as `skip()` for now till we can better understand what the underlying problem with the tests is.