Switch config stub to replace and stop reusing
In rails we commonly will create a test double, but then alter the values it uses within each test. It's a tidy feature that saves duplication.

However, spotted a comment that turns out to be from one of Sinon's maintainers saying that doing this is an anti-pattern for their framework. This was one of the reasons they added `replace()` to the framework. So this change updates the tests to use `replace()`.

https://stackoverflow.com/a/54309645/6117745