Fix code to behave as expected
With the test in place it turns out the app is not currently behaving as expected! Based on the tests it does appear to be allowing POST requests with no payloads through.

Adding the hook to the `onPostAuth` extension resolves the issue. Clearly we have some reading to do on just what is happening when with the payload during the Hapi lifecycle. But this fixes the issue we currently have.