Add additional test coverage
This adds some additional tests to confirm everyone can access the public endpoints including those without a bearer token.

We've specifically not added coverage for the condition of the authorised system user name not matching the regime ID as that's not how authorisation actually works. So we are going to focus our efforts on getting authorsation properly implemented instead.