Add tests and update Disinfect config
Doing these tests has exposed a misunderstanding about what Disinfect offers when it comes to whitespace.

> **deleteWhitespace** - remove whitespace query, payload, or params keys.

We read that as it would remove whitespace from payloads that contained properties with unnecessary whitespace. Actually, it works by removing keys where the value is _only_ whitespace.

So there is some benefit to what Disinfect offers. But we'll need to add our own `genericSanitizer()` to get the behaviour we want.