Update tests to cover nested objects
When looking into Disinfect to see how it handles nested objects, when it comes to whitespace and empty keys we spotted that unlike its sanitize method it doesn't seem to be checking the type of the property.

As suspected once we added a test that included nested objects we found these weren't being covered. Which means Disinfects functions are useless to us 😭