Add SOP data validation plugin (#37)
https://trello.com/c/xa2xIrRg

The existing [charging-module-api](https://github.com/defra/charging-module-api) checks for certain characters we've been told SOP considers invalid and can't handle.

However, it does this on an ad-hoc basis at the points we currently understand to be at risk. This has the effect of making some controllers more complex and adds duplication to the code.

On the basis our purpose is to collect and transform data before sending it onto SOP we think it's clearer to just reject any request that includes these invalid characters. There is less risk if we ensure none of our data contains them. It also simplifies the code and makes it easier to maintain.

So this change adds a custom plugin which checks for the invalid characters in any request. Unlike our previous checks of the request payload this time we won't sanitize the data. Instead, we will return a 422 response highlighting that the request contains one of the SOP invalid characters.