Fix github action ci change
https://github.com/DEFRA/sroc-charging-module-api/pull/41

When we first created the workflow file for PR #41 we incorrectly put the `ci.yml` file in the folder `.github/workflow`. It needs to be in workflows to be picked up so we used a git action to mv it.

However something went wrong because we just got a copy rather than the old one going away. This explains why we thought we had fixed the database name issue when in fact we'd amended the wrong file.

This removes the duplicate actions config file and folder.