Add back check for use of `only()` in tests to CI (#45)
https://github.com/DEFRA/charging-module-api/pull/41
https://github.com/DEFRA/charging-module-api/pull/23

In PR #23 we added in a check to our CI for temporary tags. The original description was

---

The [Hapi Lab](https://github.com/hapijs/lab) test framework comes with the ability to run specific tests by adding `.only` to the describe block. For example

```javascript
describe.only('Sanitizing requests to the API', () => {
  // ...
})
```

This is great! Just don't forget to leave them in when pushing your commits to GitHub. Having not learnt my lesson, this adds a check for use of this to the Travis build.

---

When we moved the CI to GitHub actions in PR #41 we lost the check. Having figured how to do it, this adds the check back in.