Spike regimes and authorised systems
> This change is not expected to be merged!

This is a spike to understand how to get regimes and authorised systems properly implemented. This will allow us to then get authorisation properly implemented as well, allowing us to really start building the regime based endpoints.

Because an authorised system can be linked to multiple regimes, and requires us writing to the DB for both POST and PATCH this will also allow us to understand

- how to link models together via [objection](https://github.com/vincit/objection.js)
- our standard pattern for translating requests, persisting to the db, and returning a response

As such there will be _a lot_ going on in this change. Too much in fact. So we anticipate once we understand all the changes needed will break them out into smaller PR's which will be reviewed and merged as normal.