Add ObjectCleaningService
By moving the logic currently in the plugin into a service it makes it much easier to test and means we do not have to rely putting everything through the API to actually test it.

We'll still retain some of the previous tests. But our much more detailed ones can now be done quickly directly against the service. Also means we can finally test for null values which going through the APi never allowed us to before.