Update clean payloads feature test
As we now have the ObjectCleanerService tests we don't need to low level detail tests against the API. But we still want to get a sense of what the plugin and the API is doing with a request.

Hence we have updated it to this one new test that covers all the cleaning operations.