Update authorised_systems to use conventional id
Prior to this change the ID of the authorised system record was actually the client ID generated by AWS Cognito.

We're trying to keep things as consistent as possible, so this change updates it to use a UUID as all our other tables do.

This means we need to add a new `client_id` field and use it to locate the authorised system record during the authorisation stage.