Move create auth system code to service
We've moved all the code for taking the request and creating the authorised system and put it in a service. This will mean when we start writing tests (which we will) we can test 99% of the functionality without having to fire requests at a Hapi instance. This will help speed up those tests.

When we do come to start firing requests we can more easily stub the service to provide the results we need to test the controller.

Whilst we were making changes to the controller we also update how we respond to ensure we set the status code to 201 `CREATED`.