Add new base model class (#54)
> Most of the time you want the same configuration for all models and a good pattern is to create a `BaseModel` class and inherit all your models from that. You can then add all shared configuration to `BaseModel`.

[Objection documentation](https://vincit.github.io/objection.js/guide/models.html#models)

This work comes out of the spike we have been doing in [PR #48](https://github.com/DEFRA/sroc-charging-module-api/pull/48) adding authorisation systems and sorting out authorisation in the service.

We know we need to reference other models when creating the Objection relationships. But we also know there is the issue of [require loops](https://vincit.github.io/objection.js/guide/relations.html#require-loops). A base class helps resolve this because we can set the [modelPaths property](https://vincit.github.io/objection.js/api/model/static-properties.html#static-modelpaths) and have that inherited by all sub-classes. This means we no longer have to directly require related models.

Another benefit we've found is with handling the difference in naming conventions. In databases we use snake_case. In code we use camelCase. Objection supports this through [snakeCaseMappers()](https://vincit.github.io/objection.js/recipes/snake-case-to-camel-case-conversion.html#snake-case-to-camel-case-conversion). When you set a model's `columnNameMappers()` property to use `snakeCaseMappers()` it means when you interact with the model you can use camelCase properties. It also means when we dump it to a string, for example in a response, it uses camelCase. This is important to use because it means our response style matches the style we expect in requests.

If we set the `columnNameMappers()` property in a `BaseModel` all sub-classed models will inherit this ability.

Hence this change which helps prepare us for work to come.