Remove try/catch from controller
After an [investigation into how Hapi handles errors](https://github.com/DEFRA/sroc-charging-module-api/pull/56) we now understand that any error thrown during the context of a request+response if left unhandled will automatically be picked up by Hapi and converted to a `Boom.internal()` (a 500). So we don't need to include code which catches errors and converts them to Boom ones ourselves.

This means we can remove the try/catches in our controllers!