Add JSDoc @module tag to models and services
The [@module](https://jsdoc.app/tags-module.html) tag

> [..] marks the current file as being its own module. All symbols in the file are assumed to be members of the module unless documented otherwise.

Reason for adding it is it allows us to link to these objects in other JSDoc comments. This means we can start to be more specific. For example

```
* @returns {module:AuthorisedSystemModel}
```

instead of `* @returns {Object}`.

In the subsequent changes we want to make use of this.