Update authorisation plugin
Update the plugin to use the new `AuthorisationService`` so we can start authorising users properly.

We also take advantage of the fact we have the `RegimeModel` instance so store it in the request. This means any endpoint that needs the regime no longer needs to waste time querying the database for it!