Underscores in urls? Oh my!
As we have started to migrate functionality from the [charging-module-api](https://github.com/defra/charging-module-api) we have also brought across the odd issue.

One of these is the charging-module's use of underscores in some of its endpoints. We are now [fixing the issue](https://github.com/DEFRA/charging-module-api/pull/164) there, so should it go live before this version we will not be lumbered with these 'unconventional' URL endpoints.

It's perfectly permissible to use underscores. However, it's definitely not the convention for URLs. To quote Google

> Consider using punctuation in your URLs. The URL **http://www.example.com/green-dress.html** is much more useful to us than **http://www.example.com/greendress.html**. We recommend that you use hyphens (-) instead of underscores (_) in your URLs.

There is also the 'blackboard' effect on us as a delivery team. It's for this same reason we added a [`PATCH` version of the bill run `/send` endpoint](https://github.com/DEFRA/charging-module-api/pull/156). 🧑‍🏫 😁

So as we are fixing the charging module this change brings the SROC charging module inline with it.