Implement regime authorisation fully (#57)
Currently, our authorisation mimics how authorisation should happen. It handles admin and public requests. But requests to 'regime' endpoints we mimic with a simple check that the username matches the requested regime.

What it should actually be doing is checking the requested regime against the regimes the system user is linked to.

Now we have the `RegimeModel` and `AuthorisedSystemModel` properly linked via [Objection](https://github.com/vincit/objection.js) we can do this.

This change adds a new service to handle the checking. We also needed to add a new regimes test helper to support this.

Another feature is we have to query the database for the regime as part of the checking. To avoid subsequent endpoints needing to do this we stash the `RegimeModel` instance in [request.app](https://hapi.dev/api/?v=20.0.2#-requestapp).

** Add JSDoc @module tag to models and services

The [@module](https://jsdoc.app/tags-module.html) tag

> [..] marks the current file as being its own module. All symbols in the file are assumed to be members of the module unless documented otherwise.

Reason for adding it is it allows us to link to these objects in other JSDoc comments. This means we can start to be more specific. For example

```
* @returns {module:AuthorisedSystemModel}
```

instead of `* @returns {Object}`.