Update of the Dockerfile
This update provides a

- starter for 10 for a production Dockerfile
- full support for a local development environment

We still should add a healthcheck and confirm our handling of SIGTERM calls to justify declaring this 'ready for production'. And with wanting to only install the primary dependencies a multi-stage build is likely to be the solution.

But for now this gives us a base to build on and sorts out developing with docker. With these changes

- it should work 'out of the box' after a fresh git clone
- it has the flexibility to support running both in Docker and locally
- it removes the need for additional calls/setup

That last point refers to the fact we no longer need an `ENTRYPOINT` declaration. This will be made clearer in subsequent changes. Suffice to say on first startup the environment will build and run without errors. Then after a running a [VSCode task](https://code.visualstudio.com/docs/editor/tasks) it will be usable. One more task and it will be ready to run tests.