Update db env vars to use postgres conventions
Because the names of our postgres env vars don't match the ones used by the PostgreSQL Docker image we're having to duplicate them. There is no need to stick to our own naming convention so this change updates the app to use the pg docker conventions.

This means we can remove _some_ of the duplication around env vars.