Update docker-compse.yml
With changes made to the env var names we've been able to simplify the postgres config. Now we just need to point both services at our `.env` file.

We've renamed our service to avoid seeing `sroc-charging-module-api_api` all over the place. Blurgh!

The introduction of build args and additional commands around the volumes are all part of solving the problem of building without having to run `npm install` locally first. It also is part of allowing devs to choose between running everything in Docker, or running things locally.

Finally, there is one thing that is not especially clear. As long as `POSTGRES_USER`, `POSTGRES_PASSWORD`, and `POSTGRES_DB` are provided when the postgres service is first initialised it will automatically create a DB and user with those credentials. This removes the need to run `npm run createdb` before we can even start the service.

Dealing with migrations and creating the DB can be handled by vscode tasks which will be added in a subsequent commit.