Allow and add vscode config to the project
This is a debateable change. Currently, every one on the team is using VSCode for development. VSCode tasks provide a much cleaner embedded solution to running the commands we have in the `makefile` and mangaging the terminals ourselves.

However, is it right to assume all will use them, or that future team members will be VSCode users? An alternative might be to provide an example VSCode `tasks.json` file that team members can choose to use, amend or ignore as they see fit 🤷

For now this commit adds it to help get us started using Docker. As such the subsequent README changes also assume VSCode usage for consistency. There is an argument to be made there as well instead it should use the raw docker commands and refer VSCode readers to the `tasks.json`.